implementation ther validation rule for User and Gallery

+++ app/Controllers/Http/UserController.js
@@ -16,23 +16,17 @@ class UserController {
   }
 
   async store({ request, response }) {
-    const data = request.only(['name', 'email', 'password', 'is_admin'])
-    const avatar = request.file('avatar', {
-      types: ['image'],
-      size: '10mb'
-    })
+    const data = this.getRequestData(request)
 
-    // gera um nome para a imagem do avatar
-    const avatarName = `${new Date().getTime()}.jpg`
-    if (avatar) {
-      data.avatar = avatarName
-    }
+    const { avatarImage } = data
+
+    delete data.avatarImage
 
     const model = await Model.create(data)
 
-    if (avatar) {
+    if (avatarImage) {
       // lê a imagem
-      const image = await ImageLib.readImage(avatar.tmpPath)
+      const image = await ImageLib.readImage(avatarImage.tmpPath)
       if (!image) {
         return response.status(400).json({
           message: 'Error to read avatar image.',
@@ -41,7 +35,7 @@ class UserController {
         })
       }
       // salva a imagem em um diretório
-      if (!(await this.saveAvatar(image, avatarName))) {
+      if (!(await this.saveAvatar(image, data.avatar))) {
         return response.status(500).json({
           message: 'Error to store avatar on server.',
           name: 'error',
@@ -53,42 +47,84 @@ class UserController {
     return model
   }
 
-  async show({ params }) {
+  async register({ request, response }) {
+    const data = this.getRequestData(request)
+    data.is_admin = 'false'
+
+    const { avatarImage } = data
+
+    delete data.avatarImage
+
+    const model = await Model.create(data)
+
+    if (avatarImage) {
+      // lê a imagem
+      const image = await ImageLib.readImage(avatarImage.tmpPath)
+      if (!image) {
+        return response.status(400).json({
+          message: 'Error to read avatar image.',
+          name: 'error',
+          status: 400
+        })
+      }
+      // salva a imagem em um diretório
+      if (!(await this.saveAvatar(image, data.avatar))) {
+        return response.status(500).json({
+          message: 'Error to store avatar on server.',
+          name: 'error',
+          status: 500
+        })
+      }
+    }
+
+    return model
+  }
+
+  async show({ auth, params, response }) {
+    const user = await auth.getUser()
+    if (!user.is_admin && user.id !== Number(params.id)) {
+      return response.status(401).json({
+        message: "You cannot see someone else's profile.",
+        name: 'error',
+        status: 401
+      })
+    }
     const model = await Model.findOrFail(params.id)
     return model
   }
 
-  async update({ params, request, response }) {
-    const data = request.only(['name', 'email'])
-    const avatar = request.file('avatar', {
-      types: ['image'],
-      size: '10mb'
-    })
+  async avatar({ params, response }) {
+    return response.download(`${TMP_USER}/${params.avatar}`)
+  }
 
-    // gera um nome para a imagem do avatar
-    const avatarName = `${new Date().getTime()}.jpg`
-    if (avatar) {
-      data.avatar = avatarName
+  async update({ auth, params, request, response }) {
+    const user = await auth.getUser()
+    if (!user.is_admin && user.id !== Number(params.id)) {
+      return response.status(401).json({
+        message: "You cannot update someone else's profile.",
+        name: 'error',
+        status: 401
+      })
     }
+    const data = this.getRequestData(request)
+
+    if (!user.is_admin) {
+      data.is_admin = 'false'
+    }
+    const { avatarImage } = data
+    delete data.avatarImage
 
     const model = await Model.findOrFail(params.id)
 
     // remove o avatar antigo do servidor
     await ImageLib.destroyImage(TMP_USER, model.avatar)
 
-    if (request.input('password')) {
-      data.password = request.input('password')
-    }
-    if (request.input('is_admin')) {
-      data.is_admin = request.input('is_admin')
-    }
-
     model.merge(data)
     await model.save()
 
-    if (avatar) {
+    if (avatarImage) {
       // lê a imagem
-      const image = await ImageLib.readImage(avatar.tmpPath)
+      const image = await ImageLib.readImage(avatarImage.tmpPath)
       if (!image) {
         return response.status(400).json({
           message: 'Error to read avatar image.',
@@ -97,7 +133,7 @@ class UserController {
         })
       }
       // salva a imagem em um diretório
-      if (!(await this.saveAvatar(image, avatarName))) {
+      if (!(await this.saveAvatar(image, data.avatar))) {
         return response.status(500).json({
           message: 'Error to store avatar on server.',
           name: 'error',
@@ -109,7 +145,12 @@ class UserController {
     return model
   }
 
-  async destroy({ params }) {
+  async destroy({ auth, params, response }) {
+    const user = await auth.getUser()
+    if (!user.is_admin) {
+      return response.status(401)
+    }
+
     const model = await Model.findOrFail(params.id)
     await ImageLib.destroyImage(TMP_USER, model.avatar)
     await model.delete()
@@ -131,6 +172,34 @@ class UserController {
       return false
     }
     return true
+  }
+
+  getRequestData(request) {
+    const data = {}
+    if (request.input('name')) {
+      data.name = request.input('name')
+    }
+    if (request.input('email')) {
+      data.email = request.input('email')
+    }
+    if (request.input('password')) {
+      data.password = request.input('password')
+    }
+    if (request.input('is_admin')) {
+      data.is_admin = request.input('is_admin')
+    }
+
+    const avatar = request.file('avatar', {
+      types: ['image'],
+      size: '10mb'
+    })
+    // gera um nome para a imagem do avatar
+    const avatarName = `${new Date().getTime()}.jpg`
+    if (avatar) {
+      data.avatar = avatarName
+      data.avatarImage = avatar
+    }
+    return data
   }
 }
